next hour command
