+++ apps/meeple/lib/meeple.ex
@@ -5,7 +5,11 @@ defmodule Meeple do
     send_command({:admin, :create, name: name})
   end
 
+  def next_hour() do
+    send_command({:admin, :next_hour})
+  end
+
   def clear_plan() do
-    send_command({:user, :clear_plan})
+    send_command({:admin, :clear_plan})
   end
 end
