+++ apps/meeple/lib/meeple/board.ex
@@ -5,6 +5,14 @@ defmodule Meeple.Board do
   alias Sim.Realm.SimulationLoop
   alias Meeple.{Tableau, FogOfWar, Pawn, Plan, Territory}
 
+  def create(name) do
+    :ok = Territory.create(name)
+    :ok = Tableau.create(name)
+    Tableau.pawns() |> Enum.each(&Territory.set_pawn(&1))
+    :ok = FogOfWar.create(name)
+    :ok = Plan.clear()
+  end
+
   def map_exists?() do
     Territory.exists?()
   end
