+++ apps/meeple/lib/meeple/services/user_service.ex
@@ -3,9 +3,4 @@ defmodule Meeple.Service.User do
 
   # alias Sim.Realm.SimulationLoop
   alias Meeple.Board
-
-  def execute(:clear_plan, []) do
-    Board.clear_plan()
-    [{:plan_updated, :cleared}]
-  end
 end
